hierkeys: move MasterKey conversion helpers into masterkey.go
GaloisKeyToMasterKey, MasterKeyToGaloisKey, and the automorph helpers
they use (automorphGadgetCiphertext, automorphInPlace) live next to
the MasterKey type they convert. They had no business sitting in
hierkeys.go, which is supposed to be a thin shared-utilities file.

Pure file move; no behavior or signature change.
